initial commit
An attempt to create a useful default template for bash scripts.

This template is not designed to be portable or POSIX compliant. It is
for bash scripts and follows recommendations in the Google Shell Style
Guide: google.github.io/styleguide/shell.xml

Some of the decisions I made are below:

* Using #!/usr/bin/env bash for "safety"

* error if any of the follwoing happens: a command fails, a variable is
* unset or a command in a pipechain fails

* add a trap for CTRL-C

* add some simple colors for logging - the less the better!

* Have a "main" function to call other functions. This allows us to use
* more local variables and thus increase "safety" of the script